streamlit_patch_22_aaaa

+++ streamlit_main.py
@@ -2,14 +2,15 @@ import streamlit as st
 from PIL import Image
 import tensorflow as tf
 from tensorflow_addons.layers import InstanceNormalization
-import io
+
 
 def predict_target(model, filename):
     img = tf.io.read_file(filename)
     img = tf.image.decode_image(img, channels=1)
     im = tf.cast(tf.image.resize(img, size=[32, 32]), dtype=tf.float32)
     pred = tf.squeeze(model.predict(tf.expand_dims(im, axis=0)))
-    return tf.io.encode_png(pred)
+    return tf.io.encode_png(tf.cast(pred, dtype=tf.uint16))
+
 
 def main():
     st.set_page_config(
@@ -20,7 +21,8 @@ def main():
     st.title('Deepfake demo')
     st.caption('Created using TensorFlow')
     st.write('This is demonstration of the power of the CycleGAN AKA deepfake technology.')
-    st.write('NN trained on CIFAR10 dataset provided by [TensorFlow Datasets](https://www.tensorflow.org/datasets/catalog/cifar10?authuser=3).')
+    st.write(
+        'NN trained on CIFAR10 dataset provided by [TensorFlow Datasets](https://www.tensorflow.org/datasets/catalog/cifar10?authuser=3).')
     st.write('Used dataset includes 60000 32x32 colour images in 10 classes, with 6000 images per class.')
     st.write('There are 50000 training images and 10000 test images.')
     st.subheader('Training process')
@@ -38,7 +40,8 @@ def main():
     col1.image(original_target)
     col2.image(step_target)
 
-    model = tf.keras.models.load_model('cyclegan_cifar10-g_target.h5', custom_objects={'InstanceNormalization': InstanceNormalization})
+    model = tf.keras.models.load_model('cyclegan_cifar10-g_target.h5',
+                                       custom_objects={'InstanceNormalization': InstanceNormalization})
     st.subheader('Deepfake on portraits')
     col1, col2, col3 = st.columns(3)
     turing = predict_target(model, 'portraits/alan_turing.png')
@@ -48,10 +51,7 @@ def main():
     tesla = predict_target(model, 'portraits/nikola_tesla.png')
     col3.image(tesla, use_column_width='always', caption='Nikola Tesla')
     st.subheader('Deepfake on custom images')
-    #col1, col2, col3 = st.columns(3)
-
-
-
+    # col1, col2, col3 = st.columns(3)
 
 
 if __name__ == '__main__':
